+++ DragonFrontCompanion/Controls/CardControl.xaml.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 using Xamarin.Forms;
 
-namespace DragonFrontCompanion.Contols
+namespace DragonFrontCompanion.Controls
 {
     public partial class CardControl : ContentView
     {
