+++ DragonFrontCompanion/Controls/CardCount.xaml
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="UTF-8" ?>
 <ContentView
-    x:Class="DragonFrontCompanion.Contols.CardCount"
+    x:Class="DragonFrontCompanion.Controls.CardCount"
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:ffimage="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms">
