+++ DragonFrontCompanion/Controls/CardTypeFilterLegacy.xaml
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="UTF-8" ?>
 <t:SlideMenuView
-    x:Class="DragonFrontCompanion.Contols.CardTypeFilterLegacy"
+    x:Class="DragonFrontCompanion.Controls.CardTypeFilterLegacy"
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:c="clr-namespace:DragonFrontDb.Enums;assembly=DragonFrontDb"
