+++ DragonFrontCompanion/Controls/FactionPicker.xaml.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 using Xamarin.Forms;
 
-namespace DragonFrontCompanion.Contols
+namespace DragonFrontCompanion.Controls
 {
     public partial class FactionPicker : SlideMenuView
     {
