+++ DragonFrontCompanion/Controls/ImageButton.cs
